worker/requirements.txt
mysql-connector-python
redis
minio
openai
